+++ eventbridge-20200401/alibabacloud_eventbridge20200401/models/_get_event_streaming_response_body.py
@@ -1790,6 +1790,7 @@ class GetEventStreamingResponseBodyDataSink(DaraModel):
         sink_https_parameters: main_models.SinkHttpsParameters = None,
         sink_kafka_parameters: main_models.GetEventStreamingResponseBodyDataSinkSinkKafkaParameters = None,
         sink_mnsparameters: main_models.GetEventStreamingResponseBodyDataSinkSinkMNSParameters = None,
+        sink_mqttparameters: main_models.SinkMQTTParameters = None,
         sink_ossparameters: main_models.SinkOSSParameters = None,
         sink_open_source_rabbit_mqparameters: main_models.GetEventStreamingResponseBodyDataSinkSinkOpenSourceRabbitMQParameters = None,
         sink_rabbit_mqmeta_parameters: main_models.SinkRabbitMQMetaParameters = None,
@@ -1820,6 +1821,7 @@ class GetEventStreamingResponseBodyDataSink(DaraModel):
         self.sink_kafka_parameters = sink_kafka_parameters
         # The parameters that are returned if the event target is Message Service (MNS).
         self.sink_mnsparameters = sink_mnsparameters
+        self.sink_mqttparameters = sink_mqttparameters
         self.sink_ossparameters = sink_ossparameters
         # Sink Open Source RabbitMQ Parameters
         self.sink_open_source_rabbit_mqparameters = sink_open_source_rabbit_mqparameters
@@ -1865,6 +1867,8 @@ class GetEventStreamingResponseBodyDataSink(DaraModel):
             self.sink_kafka_parameters.validate()
         if self.sink_mnsparameters:
             self.sink_mnsparameters.validate()
+        if self.sink_mqttparameters:
+            self.sink_mqttparameters.validate()
         if self.sink_ossparameters:
             self.sink_ossparameters.validate()
         if self.sink_open_source_rabbit_mqparameters:
@@ -1931,6 +1935,9 @@ class GetEventStreamingResponseBodyDataSink(DaraModel):
 
         if self.sink_mnsparameters is not None:
             result['SinkMNSParameters'] = self.sink_mnsparameters.to_map()
+
+        if self.sink_mqttparameters is not None:
+            result['SinkMQTTParameters'] = self.sink_mqttparameters.to_map()
 
         if self.sink_ossparameters is not None:
             result['SinkOSSParameters'] = self.sink_ossparameters.to_map()
@@ -2019,6 +2026,10 @@ class GetEventStreamingResponseBodyDataSink(DaraModel):
         if m.get('SinkMNSParameters') is not None:
             temp_model = main_models.GetEventStreamingResponseBodyDataSinkSinkMNSParameters()
             self.sink_mnsparameters = temp_model.from_map(m.get('SinkMNSParameters'))
+
+        if m.get('SinkMQTTParameters') is not None:
+            temp_model = main_models.SinkMQTTParameters()
+            self.sink_mqttparameters = temp_model.from_map(m.get('SinkMQTTParameters'))
 
         if m.get('SinkOSSParameters') is not None:
             temp_model = main_models.SinkOSSParameters()