+++ eventbridge-20200401/alibabacloud_eventbridge20200401/models/_update_event_streaming_request.py
@@ -1687,6 +1687,7 @@ class UpdateEventStreamingRequestSink(DaraModel):
         sink_https_parameters: main_models.SinkHttpsParameters = None,
         sink_kafka_parameters: main_models.UpdateEventStreamingRequestSinkSinkKafkaParameters = None,
         sink_mnsparameters: main_models.UpdateEventStreamingRequestSinkSinkMNSParameters = None,
+        sink_mqttparameters: main_models.SinkMQTTParameters = None,
         sink_ossparameters: main_models.SinkOSSParameters = None,
         sink_open_source_rabbit_mqparameters: main_models.UpdateEventStreamingRequestSinkSinkOpenSourceRabbitMQParameters = None,
         sink_prometheus_parameters: main_models.UpdateEventStreamingRequestSinkSinkPrometheusParameters = None,
@@ -1723,6 +1724,7 @@ class UpdateEventStreamingRequestSink(DaraModel):
         self.sink_kafka_parameters = sink_kafka_parameters
         # The parameters that are configured if you specify Simple Message Queue (SMQ, formerly MNS) as the event target.
         self.sink_mnsparameters = sink_mnsparameters
+        self.sink_mqttparameters = sink_mqttparameters
         self.sink_ossparameters = sink_ossparameters
         # The parameters that are configured if you specify open source RabbitMQ as the event target.
         self.sink_open_source_rabbit_mqparameters = sink_open_source_rabbit_mqparameters
@@ -1770,6 +1772,8 @@ class UpdateEventStreamingRequestSink(DaraModel):
             self.sink_kafka_parameters.validate()
         if self.sink_mnsparameters:
             self.sink_mnsparameters.validate()
+        if self.sink_mqttparameters:
+            self.sink_mqttparameters.validate()
         if self.sink_ossparameters:
             self.sink_ossparameters.validate()
         if self.sink_open_source_rabbit_mqparameters:
@@ -1838,6 +1842,9 @@ class UpdateEventStreamingRequestSink(DaraModel):
 
         if self.sink_mnsparameters is not None:
             result['SinkMNSParameters'] = self.sink_mnsparameters.to_map()
+
+        if self.sink_mqttparameters is not None:
+            result['SinkMQTTParameters'] = self.sink_mqttparameters.to_map()
 
         if self.sink_ossparameters is not None:
             result['SinkOSSParameters'] = self.sink_ossparameters.to_map()
@@ -1929,6 +1936,10 @@ class UpdateEventStreamingRequestSink(DaraModel):
         if m.get('SinkMNSParameters') is not None:
             temp_model = main_models.UpdateEventStreamingRequestSinkSinkMNSParameters()
             self.sink_mnsparameters = temp_model.from_map(m.get('SinkMNSParameters'))
+
+        if m.get('SinkMQTTParameters') is not None:
+            temp_model = main_models.SinkMQTTParameters()
+            self.sink_mqttparameters = temp_model.from_map(m.get('SinkMQTTParameters'))
 
         if m.get('SinkOSSParameters') is not None:
             temp_model = main_models.SinkOSSParameters()